modified ProtectedRoute to use lazy loading
